+++ tests/integration/test_api_loader.py
@@ -9,3 +9,9 @@ pytestmark = pytest.mark.django_db
 def test_load_checklists(api_key, locale, country):
     loader = APILoader(api_key, locale)
     loader.load_checklists(country, datetime.date.today())
+
+
+def test_update_checklists(api_key, locale, country):
+    loader = APILoader(api_key, locale)
+    loader.load_checklists(country, datetime.date.today())
+    loader.update_checklists(datetime.date.today())